Working Static Prototype
The Unity Scene is able to create spheres for the solar system; where each Unity scene unit is equivalent to 1 AU. All spheres are scaled accordingly. In order to render the planets and sun; their individual scale must be adjusted, but this works for a crude approximation for visualizing the size of Sagitarrius A* from various points.